+++ src/gql/base/ElementArguments.php
@@ -7,6 +7,11 @@
 
 namespace craft\gql\base;
 
+use craft\gql\types\input\criteria\Asset;
+use craft\gql\types\input\criteria\Category;
+use craft\gql\types\input\criteria\Entry;
+use craft\gql\types\input\criteria\Tag;
+use craft\gql\types\input\criteria\User;
 use craft\gql\types\QueryArgument;
 use GraphQL\Type\Definition\Type;
 
@@ -46,7 +51,7 @@ abstract class ElementArguments extends Arguments
             ],
             'siteId' => [
                 'name' => 'siteId',
-                'type' => Type::string(),
+                'type' => Type::listOf(QueryArgument::getType()),
                 'description' => 'Determines which site(s) the elements should be queried in. Defaults to the current (requested) site.'
             ],
             'unique' => [
@@ -82,7 +87,32 @@ abstract class ElementArguments extends Arguments
             'relatedTo' => [
                 'name' => 'relatedTo',
                 'type' => Type::listOf(QueryArgument::getType()),
-                'description' => 'Narrows the query results to elements that relate to *any* of the provided element IDs. This argument is ignored, if `relatedToAll` is also used.'
+                'description' => 'Narrows the query results to elements that relate to the provided element IDs. This argument is ignored, if `relatedToAll` is also used.'
+            ],
+            'relatedToAssets' => [
+                'name' => 'relatedToAssets',
+                'type' => Asset::getType(),
+                'description' => 'Narrows the query results to elements that relate to an asset list defined with this argument.'
+            ],
+            'relatedToEntries' => [
+                'name' => 'relatedToEntries',
+                'type' => Entry::getType(),
+                'description' => 'Narrows the query results to elements that relate to an entry list defined with this argument.'
+            ],
+            'relatedToUsers' => [
+                'name' => 'relatedToUsers',
+                'type' => User::getType(),
+                'description' => 'Narrows the query results to elements that relate to a use list defined with this argument.'
+            ],
+            'relatedToCategories' => [
+                'name' => 'relatedToCategories',
+                'type' => Category::getType(),
+                'description' => 'Narrows the query results to elements that relate to a category list defined with this argument.'
+            ],
+            'relatedToTags' => [
+                'name' => 'relatedToTags',
+                'type' => Tag::getType(),
+                'description' => 'Narrows the query results to elements that relate to a tag list defined with this argument.'
             ],
             'relatedToAll' => [
                 'name' => 'relatedToAll',